Allow to replace HttpSender implementation
Add two interfaces to allow to replace the implementation:
 - `HttpSenderContext`, to keep the state of each `HttpSender`
 instance;
 - `HttpSenderImpl`, creates contexts and sends the messages. It also
 keeps the listeners as they are shared between all `HttpSender`s.

Extract the current implementation into `HttpSenderContextParos` and
`HttpSenderParos`.
An add-on should now be able to set its own implementation.
